requirements.txt
pyserial>=3.5
types-pyserial>=3.5
